pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.1.2",
    "opencv-python>=4.10.0.84",
    "pillow>=10.4.0",
    "streamlit>=1.39.0",
]
